+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
-## 0.1.x, January & February 2022
+## 0.1.3, March 2022
 
-- Early releases.
+- Added `has_low_distance_stdev`.
+- Added `has_few_samples`.
+- Added `is_standardized()` function to test if a feature or regression target appears to be a Z-score.
+- Changed name of `clips()` function to `clipped()` to be more predictable (it goes with `is_clipped()`).
+
+
+## 0.1.2, February 2022
+
+- Early release.
+
+
+## 0.1.1, January 2022
+
+- Early release.